refactor: 폴더구조변경
