+++ src/features/main/components/carousels/Carousels.styles.ts
